tools/pyproject.toml
[project]
name = "soc-lab-tools"
version = "0.1.0"
description = "Python tooling for SOC home lab"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "mypy>=1.5",
    "ruff>=0.4",
    "types-requests>=2.31",
    "types-PyYAML>=6.0",
    "pip-audit>=2.7",
]

[tool.setuptools]
py-modules = ["enrich_alerts", "sigma_convert", "demo_enrich"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
